fix: dialog content clipping in mobile devices
